fix bug that causes ccp to fail if github actions = n and codevoc = n

+++ hooks/post_gen_project.py
@@ -26,4 +26,5 @@ if __name__ == "__main__":
 
     if "{{cookiecutter.codecov}}" != "y":
         remove_file("codecov.yaml")
-        remove_file(".github/workflows/validate-codecov-config.yml")
+        if "{{cookiecutter.include_github_actions}}" != "y":
+            remove_file(".github/workflows/validate-codecov-config.yml")
